perf(web): defer shell dialogs out of the initial bundle
The Help docs dialog and the upload-recovery dialog were statically
imported by the always-mounted app shell, so the Radix dialog primitive,
the help-docs search index, and the resume/discard upload paths shipped
in the entry chunk that loads on every page even though they are only
needed on user action.

Load the Help dialog lazily on first open, and gate the upload-recovery
dialog behind its scan so the dialog body only loads when resumable
uploads actually exist. The lightweight recovery scan still runs on mount.

Initial critical transfer drops from ~188 KB to ~180 KB gzipped and the
Radix dialog primitive no longer loads on a normal page visit.
